+++ lib/crypto-algorithms/aes.c
@@ -240,7 +240,7 @@ int aes_encrypt_cbc(const BYTE in[], size_t in_len, BYTE out[], const WORD key[]
 	if (in_len % AES_BLOCK_SIZE != 0)
 		return(FALSE);
 
-	blocks = in_len / AES_BLOCK_SIZE;
+	blocks = (int) in_len / AES_BLOCK_SIZE;
 
 	memcpy(buf_out, iv, AES_BLOCK_SIZE);
 
@@ -262,7 +262,7 @@ int aes_encrypt_cbc_mac(const BYTE in[], size_t in_len, BYTE out[], const WORD k
 	if (in_len % AES_BLOCK_SIZE != 0)
 		return(FALSE);
 
-	blocks = in_len / AES_BLOCK_SIZE;
+	blocks = (int) in_len / AES_BLOCK_SIZE;
 
 	memcpy(buf_out, iv, AES_BLOCK_SIZE);
